[xd|xprt] Various fixes for memory handling, ref-counting and minor cleanup.
The memory handling changes include fixes for valgrind findings
like free before de-reference, re-ordering/-assigning the information
stored in xprt atributes(E.g. xp_p2) to have it common throughout all
request types and adding initialisiation of allocated memory where needed.
The modifications for the reference counting include adjustments
of the xd object. Those include false or no increase, decrease and
wrong setting of absolute values.
Sometimes the storing of the reference and the ref-counter modification
needed to be put "closer" together in the code to visualize the relation.
The minor cleanup changes include the removal of unrequired function
parameters (E.g. oflags), the reduction of indendation by modifications of
the branching.

Signed-off-by: Swen Schillig <swen@vnet.ibm.com>